removed CSS files and styled components
